error handling improvement

+++ gdfdl-ci/00-uninstall/00-run.sh
@@ -13,10 +13,14 @@ SELF="`readlink -f $0`"
 GDFDL_BASEDIR_CI_00="`dirname ${SELF}`"
 GDFDL_BASEDIR_CI="`dirname ${GDFDL_BASEDIR_CI_00}`"
 GDFDL_BASEDIR="`dirname ${GDFDL_BASEDIR_CI}`"
-GDFDL_ENTRYWRAPPER="`find "${GDFDL_BASEDIR}/.ci" -maxdepth 1 -name *.sh`"
 
-if [[ -f "${GDFDL_ENTRYWRAPPER}" ]];
+if [ -d "${GDFDL_BASEDIR}/.ci" ]
 	then
-	bash -ex "${GDFDL_ENTRYWRAPPER}" uninstall
-	rm -rf "${GDFDL_BASEDIR}/.ci"
+	GDFDL_ENTRYWRAPPER="`find "${GDFDL_BASEDIR}/.ci" -maxdepth 1 -name *.sh`"
+
+	if [[ -f "${GDFDL_ENTRYWRAPPER}" ]];
+		then
+		bash -ex "${GDFDL_ENTRYWRAPPER}" uninstall
+		rm -rf "${GDFDL_BASEDIR}/.ci"
+	fi
 fi
